Trying to fix the inexact trunkation error

+++ src/SIP_package.jl
@@ -1368,7 +1368,7 @@ function mergers_convergence(rank, mergers_arr, my_dict, num_of_iterations, resu
 					idx_src = findfirst(rank .== levels[lev]) + 1 # computes the index of the source process (one idx up the idx of the process)
 					# new_dict, status = MPI.recv(levels[lev][idx_src], lev, comm) # receives the new dict
 					#new_dict = rec_large_data(levels[lev][idx_src], lev, comm)
-					new_dict_length = MPI.Recv(Int64, comm; source = levels[lev][idx_src], tag = lev)
+					new_dict_length = MPI.Recv(UInt32, comm; source = levels[lev][idx_src], tag = lev)
 					# we recycle the memory allotted to dict_buffer from one iteration to the next
 					resize!(new_dict_buffer, new_dict_length)
 					MPI.Recv!(new_dict_buffer, comm; source = levels[lev][idx_src], tag = lev)
@@ -1385,7 +1385,7 @@ function mergers_convergence(rank, mergers_arr, my_dict, num_of_iterations, resu
 				#my_dict = transcode(ZlibCompressor, my_dict)
 				# MPI.send(my_dict, levels[lev][idx_dst], lev, comm) # sends its dict
 				#send_large_data(my_dict, levels[lev][idx_dst], lev, comm)
-				MPI.Send(UInt64(length(my_dict)), comm; dest = levels[lev][idx_dst], tag = lev)
+				MPI.Send(UInt32(length(my_dict)), comm; dest = levels[lev][idx_dst], tag = lev)
 				MPI.Send(my_dict, comm; dest = levels[lev][idx_dst], tag = lev)
 				my_dict = nothing
 				#GC.gc()
